Fixes for trackers
